The dial starts by pointing at 50.
You could follow the instructions, but your recent required official North Pole secret entrance security training seminar taught you that the safe is actually a decoy. The actual password is the number of times the dial is left pointing at 0 after any rotation in the sequence.

For example, suppose the attached document contained the following rotations:

L68
L30
R48
L5
R60
L55
L1
L99
R14
L82
Following these rotations would cause the dial to move as follows:

The dial starts by pointing at 50.
The dial is rotated L68 to point at 82.
The dial is rotated L30 to point at 52.
The dial is rotated R48 to point at 0.
The dial is rotated L5 to point at 95.
The dial is rotated R60 to point at 55.
The dial is rotated L55 to point at 0.
The dial is rotated L1 to point at 99.
The dial is rotated L99 to point at 0.
The dial is rotated R14 to point at 14.
The dial is rotated L82 to point at 32.
Because the dial points at 0 a total of three times during this process, the password in this example is 3.

+++ 2025/01/secret-entrance/src/rotation.rs
@@ -1,27 +1,28 @@
 use regex::Regex;
 
 pub trait Spatial {
-    fn distance(&self) -> i8;
+    fn distance(&self) -> i16;
 }
 
 #[derive(PartialEq, Debug, Copy, Clone)]
 pub enum Rotation {
-    Left{ d: i8 }, Right { d: i8 }
+    Left { d: i16 },
+    Right { d: i16 }
 }
 
 impl Rotation {
     pub fn parse(s: &str) -> Self {
         let captures = Regex::new(r"^([LR])(\d+)$").unwrap().captures(s).unwrap();
         match captures.get(1).unwrap().as_str() {
-            "L" => Rotation::Left { d: captures.get(2).unwrap().as_str().parse::<i8>().unwrap() },
-            "R" => Rotation::Right { d: captures.get(2).unwrap().as_str().parse::<i8>().unwrap() },
+            "L" => Rotation::Left { d: captures.get(2).unwrap().as_str().parse::<i16>().unwrap() },
+            "R" => Rotation::Right { d: captures.get(2).unwrap().as_str().parse::<i16>().unwrap() },
             _ => panic!("Unknown rotation: {}", s)
         }
     }
 }
 
 impl Spatial for Rotation {
-    fn distance(&self) -> i8 {
+    fn distance(&self) -> i16 {
         match self {
             Rotation::Left{ d } => -*d,
             Rotation::Right{ d } => *d,
